Fix baselines fail on first run due to dependancy
Currently the baselines fail on first run for a new account with the
error -

```
Error: Creating/Updating CloudWatch Log Metric Filter failed: ResourceNotFoundException: The specified log group does not exist.

  with module.baselines.module.securityhub-alarms.aws_cloudwatch_log_metric_filter.unauthorised-api-calls,
  on .terraform/modules/baselines/modules/securityhub-alarms/main.tf line 54, in resource "aws_cloudwatch_log_metric_filter" "unauthorised-api-calls":
  54: resource "aws_cloudwatch_log_metric_filter" "unauthorised-api-calls" {
```

This fails because the log group "cloudtrail" has not yet been created.
When creating the log group the only output available is ARN, and the
metric filter needs the log group name so this has been hard coded. It
is possible to manipulate the ARN to get the name, but this adds
complexity to the code just to achieve the dependancy. The simpler way
is adding a depends_on.