Fix cloudtrail sns topic policy dependancy
We currently get the following error message on first run of the
baseline:
```
Error: Error creating CloudTrail: InsufficientSnsTopicPolicyException: SNS Topic does not exist or the topic policy is incorrect!

  with module.baselines.module.cloudtrail.aws_cloudtrail.cloudtrail,
  on .terraform/modules/baselines/modules/cloudtrail/main.tf line 11, in resource "aws_cloudtrail" "cloudtrail":
  11: resource "aws_cloudtrail" "cloudtrail" {
```

I believe this is because the policy has not yet been attached to the
SNS topic, this change adds an explicit dependancy.