Remove controls around MFA
Controls around hardware MFA and MFA for root users is not needed as we
have implemented SCPs which prevent the root user from performing any
actions when logged in.

Also disabling the MFA metric and alarm control as SSO users are using MFA with
the identity provider, but this is not recognised by security hub
currently. The alarm is adjusted to ignore SSO users without MFA.